requirements.txt
fitz==0.0.1.dev2
langchain==0.3.19
langchain_community==0.3.18
langchain_core==0.3.40
langchain_gigachat==0.3.4
numpy==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
streamlit==1.42.2
